requirements.txt
yfinance
numpy
pandas
matplotlib.pyplot